feat: 마이페이지에 내가 작성한 목록 추가 및 관련 레포지토리 수정 및 생성 / static에 정적 html들 삭제함

+++ src/main/java/com/booklion/model/entity/Questions.java
@@ -33,7 +33,7 @@ public class Questions {
 	
 	@Id
 	@GeneratedValue(strategy=GenerationType.IDENTITY)
-	private Integer questId;
+	private Long questId;
 	
 	private String title;
 	private String content;
@@ -75,7 +75,7 @@ public class Questions {
 
 
 
-	public Integer getQuestId() {
+	public Long getQuestId() {
 		return questId;
 	}
 
@@ -110,7 +110,6 @@ public class Questions {
 	public void setLikeCount(int likeCount) {
 	    this.likeCount = likeCount;
 	}
-
 
 	
 }
